requirements.txt
# flask
click==7.1.1
flask-sqlalchemy==2.4.1
flask==1.1.2
itsdangerous==1.1.0
jinja2==2.11.1
markupsafe==1.1.1
python-dotenv==0.12.0
sqlalchemy==1.3.15
werkzeug==1.0.1
bootstrap-flask
PyMySQL
# aliyun sdk
aliyun-python-sdk-rds
aliyun-python-sdk-ecs
aliyun-python-sdk-core
six
aliyun-python-sdk-bssopenapi 
aliyun-python-sdk-vpc 
aliyun-python-sdk-cms 
aliyun-python-sdk-waf-openapi
aliyun-python-sdk-cdn
# oracle cloud api
oci
# web server
gunicorn
supervisor
# async
celery
redis
